backend/requirements.txt
aniso8601==3.0.0
autopep8==1.3.5
click==6.7
Flask==1.0.2
Flask-REST-JSONAPI==0.19.1
Flask-RESTful==0.3.6
flask-restplus==0.11.0
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
marshmallow==2.13.1
marshmallow-jsonapi==0.20.0
pycodestyle==2.4.0
PyJWT==1.6.4
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.8
Werkzeug==0.14.1
